chore(db): change `card.parentId` to `bigint`

+++ packages/backend/src/migrations/1720077323722-add-parent-and-children-columns-to-cards-table.ts
@@ -14,7 +14,7 @@ export class AddParentAndChildrenColumnsToCardsTable1720077323722
             "card",
             new TableColumn({
                 name: "parentId",
-                type: "int",
+                type: "bigint",
                 isNullable: true,
             })
         );
